Add files via upload
When running the code, please use allElectionData.txt. All of the functions work to my knowledge but please let me know if you do find a bug. When calling option 10, compare the expected and actual relative frequencies. Large deviations from the expected relative frequencies is often a sign of data manipulation. However, when running the Benford's test on States versus The US itself, you'll find that there is much deviation when running the test the on states because the data set is much smaller. In the 2020 election, many conservatives have tried to claim fraud in several key states using the results from a Benford's Law test on a given state's election data as proof. However, running this test on such a small data set produces highly skewed results. I've included the ability to run the test on states for learning-purposes but is otherwise useless in a real-world context.